bot/requirements.txt
langchain==0.1.3
python-dotenv==1.0.1
streamlit==1.30.0
openai==1.9.0
## The following requirements were added by pip freeze:
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
fastapi==0.109.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain-cli==0.0.21
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
langserve==0.0.39
langsmith==0.0.83
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
orjson==3.9.12
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
sse-starlette==1.8.2
starlette==0.35.1
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
yarl==1.9.4
zipp==3.17.0